requirements.txt
flask>=3.0
requests>=2.31
pycryptodome>=3.20
psycopg2-binary>=2.9
